Backend slot widths: u128 slots when the design outgrows 64 bits
The simulator is generic over a Slot type (u64 default, u128 wide —
register-pair native on 64-bit CPUs, no software emulation). --slot auto
picks wide slots only for designs with >64-bit signals; 64/128 force it.
Samples/VCD widen to u128. Floats stay f64: no scalar f128/f256 hardware
exists on mainstream CPUs, so wider floats are deferred.

Also: !=, <= and >= are now lexed and parsed (the comparison family was
incomplete — only ==, <, > worked).

+++ crates/siox-wave/src/lib.rs
@@ -37,10 +37,10 @@ pub fn write_vcd<W: Write>(out: &mut W, design: &Design, samples: &[Sample]) -> 
 
     // Value changes over time. Only emit signals whose value actually changed,
     // and one `#time` marker per distinct time.
-    let mut last: Vec<Option<u64>> = vec![None; design.signals.len()];
+    let mut last: Vec<Option<u128>> = vec![None; design.signals.len()];
     let mut cur_time: Option<u64> = None;
     for sample in samples {
-        let changes: Vec<(usize, u64)> = sample
+        let changes: Vec<(usize, u128)> = sample
             .values
             .iter()
             .enumerate()
@@ -76,7 +76,7 @@ fn vcd_width(width: u32) -> u32 {
     }
 }
 
-fn write_value<W: Write>(out: &mut W, value: u64, width: u32, id: &str) -> io::Result<()> {
+fn write_value<W: Write>(out: &mut W, value: u128, width: u32, id: &str) -> io::Result<()> {
     if width == 1 {
         writeln!(out, "{}{id}", value & 1)
     } else {
